Update the Kotlin attributes schema
First, introduce new Usage attribute values, `kotlin-api` and
`kotlin-runtime` for Kotlin non-JVM variants. We need this to be able
to publish a Kotlin library with several variants and consume it in a
pure Java project that is unaware of Kotlin. Our initial scheme would
cause a conflict of multiple `java-api` and/or `java-runtime-jars`
variants generated by multiple Kotlin targets. We should publish only
JVM artifacts with the Java Usage values.

Second, setup the compatibility rules for Kotlin platform type so that
the input configurations for Kotlin metadata can consume platform
artifacts. We need that because these configurations extend those for
normal dependencies, and if a dependency has no Kotlin metadata, the
configuration would fail to resolve.